+++ lib/gui/app.js
@@ -43,14 +43,13 @@ const updateNotifier = require('./components/update-notifier');
 const availableDrives = require('./models/available-drives');
 const selectionState = require('./models/selection-state');
 const driveScanner = require('./modules/drive-scanner');
+const osDialog = require('./os/dialog');
+const exception = require('./modules/exception');
 
 const app = angular.module('Etcher', [
   require('angular-ui-router'),
   require('angular-ui-bootstrap'),
   require('angular-if-state'),
-
-  // Etcher modules
-  require('./modules/error'),
 
   // Components
   require('./components/svg-icon'),
@@ -65,7 +64,6 @@ const app = angular.module('Etcher', [
   // OS
   require('./os/open-external/open-external'),
   require('./os/dropzone/dropzone'),
-  require('./os/dialog/dialog'),
 
   // Utils
   require('./utils/manifest-bind/manifest-bind')
@@ -85,7 +83,7 @@ app.run(() => {
   ].join('\n'));
 });
 
-app.run((ErrorService) => {
+app.run(() => {
   analytics.logEvent('Application start');
   settings.load();
 
@@ -148,7 +146,7 @@ app.run((ErrorService) => {
     return updateNotifier.notify(latestVersion, {
       allowSleepUpdateCheck: currentReleaseType === release.RELEASE_TYPE.PRODUCTION
     });
-  }).catch(ErrorService.reportException);
+  }).catch(exception.report);
 });
 
 app.run(() => {
@@ -176,7 +174,7 @@ app.run(() => {
   });
 });
 
-app.run(($timeout, ErrorService) => {
+app.run(($timeout) => {
   driveScanner.on('drives', (drives) => {
 
     // Safely trigger a digest cycle.
@@ -196,13 +194,13 @@ app.run(($timeout, ErrorService) => {
     // spamming our error reporting service.
     driveScanner.stop();
 
-    return ErrorService.reportException(error);
+    return exception.report(error);
   });
 
   driveScanner.start();
 });
 
-app.run(($window, WarningModalService, ErrorService, OSDialogService) => {
+app.run(($window) => {
   let popupExists = false;
 
   $window.addEventListener('beforeunload', (event) => {
@@ -221,7 +219,7 @@ app.run(($window, WarningModalService, ErrorService, OSDialogService) => {
 
     analytics.logEvent('Close attempt while flashing');
 
-    OSDialogService.showWarning({
+    osDialog.showWarning({
       confirmationLabel: 'Yes, quit',
       rejectionLabel: 'Cancel',
       title: 'Are you sure you want to close Etcher?',
@@ -240,7 +238,7 @@ app.run(($window, WarningModalService, ErrorService, OSDialogService) => {
 
       analytics.logEvent('Close rejected while flashing');
       popupExists = false;
-    }).catch(ErrorService.reportException);
+    }).catch(exception.report);
   });
 });
 
@@ -264,11 +262,10 @@ app.config(($urlRouterProvider) => {
 });
 
 app.config(($provide) => {
-  $provide.decorator('$exceptionHandler', ($delegate, $injector) => {
-    return (exception, cause) => {
-      const ErrorService = $injector.get('ErrorService');
-      ErrorService.reportException(exception);
-      $delegate(exception, cause);
+  $provide.decorator('$exceptionHandler', ($delegate) => {
+    return (except, cause) => {
+      exception.report(except);
+      $delegate(except, cause);
     };
   });
 });
