--- lib/gui/modules/error.js
@@ -1,55 +1,0 @@
-/*
- * Copyright 2016 resin.io
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-'use strict';
-
-/**
- * @module Etcher.Modules.Error
- */
-
-const angular = require('angular');
-const _ = require('lodash');
-const analytics = require('../modules/analytics');
-
-const MODULE_NAME = 'Etcher.Modules.Error';
-const error = angular.module(MODULE_NAME, [
-  require('../os/dialog/dialog')
-]);
-
-error.service('ErrorService', function(OSDialogService) {
-
-  /**
-   * @summary Report an exception
-   * @function
-   * @public
-   *
-   * @param {Error} exception - exception
-   *
-   * @example
-   * ErrorService.reportException(new Error('Something happened'));
-   */
-  this.reportException = (exception) => {
-    if (_.isUndefined(exception)) {
-      return;
-    }
-
-    OSDialogService.showError(exception);
-    analytics.logException(exception);
-  };
-
-});
-
-module.exports = MODULE_NAME;
