+++ lib/gui/modules/exception.js
@@ -16,26 +16,25 @@
 
 'use strict';
 
+const _ = require('lodash');
+const analytics = require('../modules/analytics');
+const osDialog = require('../os/dialog');
+
 /**
- * @module Etcher.Pages.Settings
+ * @summary Report an exception
+ * @function
+ * @public
+ *
+ * @param {Error} exception - exception
+ *
+ * @example
+ * exception.report(new Error('Something happened'));
  */
+exports.report = (exception) => {
+  if (_.isUndefined(exception)) {
+    return;
+  }
 
-const angular = require('angular');
-const MODULE_NAME = 'Etcher.Pages.Settings';
-const SettingsPage = angular.module(MODULE_NAME, [
-  require('angular-ui-router'),
-  require('../../components/warning-modal/warning-modal')
-]);
-
-SettingsPage.controller('SettingsController', require('./controllers/settings'));
-
-SettingsPage.config(($stateProvider) => {
-  $stateProvider
-    .state('settings', {
-      url: '/settings',
-      controller: 'SettingsController as settings',
-      templateUrl: './pages/settings/templates/settings.tpl.html'
-    });
-});
-
-module.exports = MODULE_NAME;
+  osDialog.showError(exception);
+  analytics.logException(exception);
+};
