+++ lib/gui/pages/main/controllers/drive-selection.js
@@ -19,8 +19,9 @@
 const settings = require('../../../models/settings');
 const selectionState = require('../../../models/selection-state');
 const analytics = require('../../../modules/analytics');
+const exception = require('../../../modules/exception');
 
-module.exports = function(ErrorService, DriveSelectorService) {
+module.exports = function(DriveSelectorService) {
 
   /**
    * @summary Open drive selector
@@ -42,7 +43,7 @@ module.exports = function(ErrorService, DriveSelectorService) {
         device: drive.device,
         unsafeMode: settings.get('unsafeMode')
       });
-    }).catch(ErrorService.reportException);
+    }).catch(exception.report);
   };
 
   /**
