+++ lib/gui/pages/main/controllers/flash.js
@@ -27,8 +27,7 @@ const path = require('path');
 module.exports = function(
   $state,
   ImageWriterService,
-  FlashErrorModalService,
-  ErrorService
+  FlashErrorModalService
 ) {
 
   /**
@@ -96,7 +95,7 @@ module.exports = function(
         FlashErrorModalService.show(messages.error.notEnoughSpaceInDrive());
       } else {
         FlashErrorModalService.show(messages.error.genericFlashError());
-        ErrorService.reportException(error);
+        error.reportException(error);
       }
 
     })
