+++ lib/gui/pages/main/controllers/image-selection.js
@@ -25,10 +25,11 @@ const imageStream = require('../../../../image-stream');
 const supportedFormats = require('../../../../shared/supported-formats');
 const analytics = require('../../../modules/analytics');
 const selectionState = require('../../../models/selection-state');
+const osDialog = require('../../../os/dialog');
+const exception = require('../../../modules/exception');
 
 module.exports = function(
-  ErrorService,
-  OSDialogService,
+  $timeout,
   WarningModalService
 ) {
 
@@ -63,7 +64,7 @@ module.exports = function(
    * @param {Object} image - image
    *
    * @example
-   * OSDialogService.selectImage()
+   * osDialogService.selectImage()
    *   .then(ImageSelectionController.selectImage);
    */
   this.selectImage = (image) => {
@@ -75,7 +76,7 @@ module.exports = function(
         })
       });
 
-      OSDialogService.showError(invalidImageError);
+      osDialog.showError(invalidImageError);
       analytics.logEvent('Invalid image', image);
       return;
     }
@@ -117,7 +118,7 @@ module.exports = function(
       image.bmap = Boolean(image.bmap);
 
       return analytics.logEvent('Select image', image);
-    }).catch(ErrorService.reportException);
+    }).catch(exception.report);
   };
 
   /**
@@ -158,17 +159,19 @@ module.exports = function(
   this.openImageSelector = () => {
     analytics.logEvent('Open image selector');
 
-    OSDialogService.selectImage().then((imagePath) => {
+    osDialog.selectImage().then((imagePath) => {
+      $timeout(() => {
 
-      // Avoid analytics and selection state changes
-      // if no file was resolved from the dialog.
-      if (!imagePath) {
-        analytics.logEvent('Image selector closed');
-        return;
-      }
+        // Avoid analytics and selection state changes
+        // if no file was resolved from the dialog.
+        if (!imagePath) {
+          analytics.logEvent('Image selector closed');
+          return;
+        }
 
-      this.selectImageByPath(imagePath);
-    }).catch(ErrorService.reportException);
+        this.selectImageByPath(imagePath);
+      });
+    }).catch(exception.report);
   };
 
   /**
