+++ lib/gui/pages/main/controllers/main.js
@@ -19,12 +19,12 @@
 const settings = require('../../../models/settings');
 const flashState = require('../../../models/flash-state');
 const analytics = require('../../../modules/analytics');
+const exception = require('../../../modules/exception');
 const availableDrives = require('../../../models/available-drives');
 const selectionState = require('../../../models/selection-state');
 
 module.exports = function(
   TooltipModalService,
-  ErrorService,
   OSOpenExternalService
 ) {
 
@@ -85,7 +85,7 @@ module.exports = function(
     return TooltipModalService.show({
       title: 'Image File Name',
       message: selectionState.getImagePath()
-    }).catch(ErrorService.reportException);
+    }).catch(exception.report);
   };
 
 };
