+++ lib/gui/pages/main/main.js
@@ -39,12 +39,10 @@ const MainPage = angular.module(MODULE_NAME, [
   require('../../components/flash-error-modal/flash-error-modal'),
   require('../../components/progress-button/progress-button'),
 
-  require('../../os/dialog/dialog'),
   require('../../os/open-external/open-external'),
   require('../../os/dropzone/dropzone'),
 
   require('../../modules/image-writer'),
-  require('../../modules/error'),
 
   require('../../utils/byte-size/byte-size')
 ]);
