+++ lib/gui/pages/settings/controllers/settings.js
@@ -20,8 +20,9 @@ const os = require('os');
 const _ = require('lodash');
 const settings = require('../../../models/settings');
 const analytics = require('../../../modules/analytics');
+const exception = require('../../../modules/exception');
 
-module.exports = function(WarningModalService, ErrorService) {
+module.exports = function(WarningModalService) {
 
   /**
    * @summary Client platform
@@ -100,7 +101,7 @@ module.exports = function(WarningModalService, ErrorService) {
         this.model.set(setting, true);
         this.refreshSettings();
       }
-    }).catch(ErrorService.reportException);
+    }).catch(exception.report);
   };
 
 };
